+++ android/app/src/main/java/app/priceguard/ui/home/list/ProductListFragment.kt
@@ -15,9 +15,9 @@ import app.priceguard.databinding.FragmentProductListBinding
 import app.priceguard.ui.additem.AddItemActivity
 import app.priceguard.ui.home.ProductSummaryAdapter
 import app.priceguard.ui.home.list.ProductListViewModel.ProductListEvent
-import app.priceguard.ui.util.drawable.disableAppBarRecyclerView
-import app.priceguard.ui.util.drawable.showNetworkDialog
 import app.priceguard.ui.util.lifecycle.repeatOnStarted
+import app.priceguard.ui.util.ui.disableAppBarRecyclerView
+import app.priceguard.ui.util.ui.showNetworkDialog
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 import kotlinx.coroutines.launch
